Solution 2
